fix: resolve next.js build errors
